Bug#32809268 sd_notify test-failure on slow machines
Issue
=====

NotifyTest.NotifyReadyMetadataCacheNoServer fails on slow machines with
"signal 15".

The test does not synchronize with the router startup, but expects a
result in 500ms which is reasonable on development machines, but isn't
enough for loaded machines.

After 500ms the test tries to shutdown, but the signal handler isn't
started yet which leads to a "terminate on signal 15".

Most of the time is spent on spawning as many io-threads as the machine
as CPU-threads, even though no io-thread is used for this test.

Change
======

- speed up router startup by starting only 1 io-thread (implicitly
  through config_writer()'s default values)
- increase timeout for NotifyReadyMetadataCacheNoServer for router to
  spawn

RB: 26291