Bug#29634540 "DROP DATABASE OF 1 MILLION TABLES RESULTED CRASH OF MYSQLD" [noclose]
Attempt to drop database with 1 million tables results in
out-of-memory condition.

With transition to New Data-Dictionary and atomic DDL, DROP DATABASE
started to allocate much more memory which is released at the end of
statement. This memory allocated for various DD API objects, new MDL
locks, temporary objects used during updates to DD tables and so on.

Part of the problem was that code responsible for dropping routines
and events in the database acquired uncached DD objects for all
routines/events to be dropped at once and then unnecessarily kept
them around until the statement end.

This patch solves the problem by avoiding constructing uncached DD
objects for routines/events to be dropped. Instead we now get names
of all routines and events in the database and then process them
one-by-one by acquiring cached DD object and then using it to drop item
being processed (the cached object is released in the process).
Such approach is similar to one that DROP DATABASE uses for tables.

There is no MTR test case as it is hard to write robust one for this
situation.

This is the next step towards solving the original problem to the full
extent.

Reviewed-by: Dyre Tjeldvoll <Dyre.Tjeldoll@oracle.com>