Bug#32406875 NON PARALLEL TESTS RUNNING IN PARALLEL IN MTR
Problem: when scheduling a test, MTR tries to be clever and schedule a
certain number of tests that have the same startup options in the same
worker thread.

However non-parallel tests can also be scheduled this way, even if
they are supposed to be run by the last active worker thread. Hence
they end up being run in parallel with other tests.

Solution: never let a worker thread pick up a non-parallel test.

Change-Id: Ib3cf8a5c290a9cfee924a0f918df0dbe04e28747