Bug#32406875 NON PARALLEL TESTS RUNNING IN PARALLEL IN MTR
This patch solves two problems in MTR's scheduling of non-parallel
tests.

Problem 1: there is a flaw in the logic where a worker thread that has
a sanitizer failure can be registered as having completed twice. This
again leads to a counting error and we end up with two or more workers
executing non-parallel tests in parallel.

Solution: make sure that workers that have received a 'BYE' message
from the master thread, never get tests scheduled again.

Problem 2: when scheduling a test, MTR tries to be clever and schedule a
certain number of tests that have the same startup options in the same
worker thread.

However non-parallel tests can also be scheduled this way, even if
they are supposed to be run by the last active worker thread. Hence
they can potentially end up being run in parallel with other tests.

Solution: never let a worker thread pick up a non-parallel test.

Change-Id: I1fd919f357d1aaa025f02e0d48af43c85425fbcf