Bug#32730214 LEAK OF MYSQL LIB RESOURCES IN DBUTIL
The DbUtil class uses the MySQL client library to connect to a MySQL Server,
this is done by initializing a MYSQL instance. Behind the scenes this
initializes the MySQL client library and allocates resource in thread local
storage of the thread using Dbutil. This is most easily seen with debug
compiled binaries that allocates some memory. The MySQL client library
resources should be released by calling mysql_library_end() when everything
in the binary has finished using the library. The thread local resources
should be released by calling mysql_thread_end() when the thread has finished
using the library.

Add automatic cleanup of MySQL library resources when program(s) using DbUtil
exits. The MySQL library end function is safe to call multiple times.
Add function for cleanup of thread local resources and call it from users of
DbUtil that are ran with memory validation tools, also this function is safe
to call multiple times but need to be done before thread exits and is not
thread safe. Fix unrelated anomalies.

Change-Id: Ib2712f6c2dfe3f61f35dbf8b7a42a0b20ff03583