Bug#32713166 NDB : MYSQLD LOGS TOO MANY ERRORS WHILE DISCONNECTED FROM DATA NODES
Returning an error while determining the number of partitions used by a
NDB table will trigger the MySQL Server to write a message about
"Incorrect information in <table>.frm" file to its error log. That's not
a correct message as there is no incorrect information in any such file
(if it would even exist). Also, writing to the error log will cause it
to flood-fill when users are trying to open tables while MySQL Server is
not connected to NDB.

Fix by changing the function that determines number of partitions to use
the value loaded from DD without fetching it from NDB, this also saves
one roundtrip during open table. For the special case where table is
opened for upgrade, fall back to fetch value from NDB in the upgrade
code path. Move misplaced function comment and also a function in the
middle of related code.

No testcase provided as both the "open table while not connected to NDB"
as well as "fill number of partitions from NDB during upgrade" are
already tested by existing tests. Rigorous testing has been performed in
order to double check that the "Incorrect ... *.frm" message is not
writen to error log.

Change-Id: I2405181f4de77163d1c9ff7be81309d2d03074a7