Bug#32712389 FIXES THE PROPERTY TO RESTART MGMDS IN EXPECTED MANNER
Modifies testUpgrade.cpp to fix the case sensitive property name,
'InitialMgmdRestart' for some test cases.

InitialMgmdRestart is a property which acts as a control variable
in deciding the type of restart to be performed on management nodes
for upgrading/downgrading the cluster. For some test cases, it was
being incorrectly set(InitialMGMDRestart) which led to Upgrade_NR1
and Upgrade_Api_Before_NR1 to restart mgmds without --initial as
opposed to with --initial which is the expected behaviour.

Change-Id: I8c85546f3bb56ac84c218c279455094a1398dc8a