Bug #32745293 MSVC DOES NOT HANDLE THROWING FROM ALLOCATOR IN DEBUG BUILDS
Some of the TempTable unit-tests, in order to test the edge cases, are
intentionally throwing an exception from TempTable::Allocator and that
somehow causes MSVC in debug builds to generate such code which makes
that exception to propagate all the way down to the C++ runtime resulting
with a crash.

Interestingly this issue is not visible in MSVC release builds, and GCC
and clang can happily handle this code both in debug and release builds.

Issue was reported to Microsoft but was closed as not a bug with the
explanation: "This is another instance of the 'iterator debugging
doesn’t support handling out of memory situations' bit."

Current trunk disables such unit-tests from running on all debug builds.
In order to gain somewhat better test coverage, this patch modifies the
code to exclude those tests on Windows debug builds only.

Reviewed-by: Tor Didriksen <tor.didriksen@oracle.com>