Bug #32746522 UT::ALIGNED_* UNIT-TEST ASSERT(PFS_METADATA_SIZE <= ALIGNMENT) ON WINDOWS
Under Windows, ut::aligned_* suite of unit-tests triggers the debug
assertion in the following part of ut::aligned_* implementation:

  static inline void pfs_metaoffset(
      Aligned_alloc_impl::data_segment_ptr data,
      std::size_t alignment) noexcept {
    ...
    assert(pfs_metadata_size <= alignment);
    ...
  }

Under Linux, and other tested platforms, same unit-test runs fine. Other
than observed assertion in unit-test code, there are no other consequences,
e.g. no crashes or whatsoever are observed on Windows test-runs.

However, unit-test under subject revealed an actual bug in
ut::aligned_* implementation which is only relevant to Windows implementation
and which may occur on Windows under very specific circumstances.

Windows 32-bit and 64-bit builds both have alignof(max_align_t) set to
8 (bytes), and cross-platform ut::aligned_* implementation was wrongly
exploiting this detail to create an extra buffer of at least
2 * alignof(max_align_t) bytes large to hold the PFS metadata.
Given that PFS metadata (in 64-bit builds) will be as large as 24 bytes,
2 * alignof(max_align_t) large buffer (16 bytes on Windows) will
obviously not be enough to accommodate the PFS metadata and that will
lead to the assert under subject.

Interestingly, condition to reach the assert occurs on Windows under
single circumstance and that is when user code requests the
ut::aligned_* library to dynamically allocate data that is overaligned
to 16 bytes. Otherwise, asking the ut::* library to allocate and
overalign the data to any other value larger than 16 will work just fine.

 ut::aligned_alloc(n_bytes, alignment=16);   -> assertion (bug on Windows)
 ut::aligned_alloc(n_bytes, alignment=32);   -> OK
 ut::aligned_alloc(n_bytes, alignment=64);   -> OK
 ut::aligned_alloc(n_bytes, alignment=128);  -> OK
 ut::aligned_alloc(n_bytes, alignment=256);  -> OK
 ...
 ut::aligned_alloc(n_bytes, alignment=4096); -> OK
 ...
 ut::aligned_alloc(n_bytes, alignment=N);    -> OK

To fix the issue, and to keep the implementation bug-free across platforms,
special care must be taken to allocate enough extra space to hold the
PFS metadata but we also need to make sure that the pointer which will be
returned to the callee by this ut::aligned_* will still be suitably
over-aligned as the user code requested.

Both of these requirements can be fulfilled by finding the smallest
multiple of requested alignment that is not smaller than actual PFS
metadata size.

Nice property of such transformation is that the code will not be asking
no more memory than it is really needed and, given that ut::aligned_* is
operating with 2^N numbers, computing such a transformation on x86 can
be implemented at the negligible cost of single negation and a bitwise AND.

Change-Id: I84d3a0d568069cc606166ee54acd51ec985edf96