BUG#32668567: VIEW_CHANGE_LOG_EVENT DOES NOT SET COMMIT TIMESTAMPS
Description:
WL#7319: "Infrastructure for GTID based delayed replication and replication
lag monitoring" specifies:
I2. A timestamp - the *original_commit_timestamp* - shall be generated when a
transaction is committed on the first master, stored in the
Gtid_log_event in the binary log, and propagated to binary logs of slaves
using the @@session.original_commit_timestamp variable. For a replicated
transaction, the original_commit_timestamp must be identical in the
SOURCE binary log and in all REPLICA binary logs.

If the transaction was generated by a group member, the
original_commit_timestamp shall be generated when the transaction is
ready to be committed.

Transactions that consist of a View_change_log_event are the exception:
this event is created locally on each group member every time a member
joins and it is not replicated.
The original_commit_timestamp is set to zero as there is no single
original source.
This means that for View_change original_commit_timestamp and
immediate_commit_timestamp are 0.

This has no impact on data correctness, though it damages observability,

Resolution:
GR will now generate both 'immediate_commit_timestamp' and
'original_commit_timestamp' for transactions associated to
View Change Log Events.
Both timestamps must be equal per server, but each member that generates
this transaction will log it with local timestamp values.

ReviewBoard: 26127