WL #14418: Support for cyclic hypergraphs [patch 16/16, delay concretization]
Delay concretizing multiple equalities in join conditions for semijoins
and antijoins until we know for sure what tables are referred to by other
join conditions.

This fixes a specific issue in DBT-3 Q21, where the antijoins and
semijoin would come in the wrong order, due to inability to pick the
best concretization of the multiple equality in the presence of another
(non-equijoin) condition. However, it should be applicable across all
such (rare) queries.

Change-Id: I9bb8a13bcb531e029282bc3396112162b334643a