Bug #32762594: FSM FOR INTERESTING ORDERS GETS TOO LARGE
select.inc has a query that sets up an 8-way join with a 5-element interesting
grouping, where all fields can be turned into all fields by way of a large
multi-equality, and all fields have a usable index. This causes the state space
in the FSM for interesting orders to explode, since nearly any permutation of
fields is reachable and can lead to the grouping; it produces a 8821-state NFSM
(none of which gets pruned away in the pruning step, short of 31 edges),
which would probably require roughly 2^50 DFSM states to represent after
conversion. Needless to say, this is not practically realizable, and the query
times out.

We add some fairly generous limits of max. 200 NFSM states (before pruning)
and 2000 DFSM states; after that, we simply stop expanding the states.
(We keep the information we already have, which means we can often still do
limited sortahead or the likes; we just cannot follow functional dependencies.)
This is enough for most practical queries that we can solve at all, but when
hitting the limits, it gives up in about 30 ms on a fast CPU.

Change-Id: I210fac290ceb71db421dcc049ec45c4523bba9cb