Bug#32104924: INNODB_UNDO.UNDO_DDL_RECOVER FAILS ON PB2
There is a race condition between the purge thread that is truncating an
undo space and a thread that does a select from information_schema.files.

The purge thread calls trx_purge_truncate_marked_undo() where it will get
and hold;
  1) an MDL on the space name and
  2) the undo::ddl_mutex
While holding those, it calls trx_purge_truncate_marked_undo_low().
This routine gets and releases undo::spaces->s_lock() and then the x_lock()
while marking the rsegs of the undo::Tablespace as 'empty' or 'active'
depending on whether this was an explicit truncation or an implicit one.

The next thing that trx_purge_truncate_marked_undo_low() does is update
DD:tablespace::state in a call to dd_tablespace_set_id_and_state().
This is a DD transaction which should prevent any other transaction
from reading this change until it is committed. Once it is committed,
the state change to 'empty' should be readable to other transactions.

Many testcases use mysql-test/include/wait_until_undo_space_is_empty.inc
which queries INFORMATION_SCHEMA.FILES in a loop so that as soon as the
undo space is empty, it can move on.  Then they often call DROP UNDO
TABLESPACE.  The testcase innodb_undo.undo_ddl_recover immediately tries
to display this empty state in a SELECT FROM INFORMATION_SCHEMA.FILES.
This fails to find the empty file sporadically on PB2 and when running
this test a large number of times in parallel.

The problem is that SELECTs from INFORMATION_SCHEMA.FILES use MVCC to
see the visible version of records selected from dd::tablespaces and
dd::tablespace_files. But when selecting records from INFORMATION_SCHEMA.
INNODB_TBALESPACES, MVCC is not used.

So the race condition happens when wait_until_undo_space_is_empty.inc
returns as soon as it reads 'state=empty' from i_s.innodb_tablespaces
and then that transaction is not committed until after the following
query for I_S.FILES is started.  The I_S.FILES query is looking at
the old space ID which has already been taken out of the InnoDB search
list.  It is still accessible for pages from that file still in the
buffer pool but it is no longer found when searching by space_id.

So in the rare case that dd_tablespace_set_id_and_state() makes a new
record version in dd::tablespaces for the new undo tablespace, and that
record is read immediately by wait_until_undo_space_is_empty.inc querying
I_S.INNODB_TABLESPACES, then the testcase starts a query to I_S.FILES
before the purge thread in dd_tablespace_set_id_and_state() can commit
the change.  Then the output of the test will be different when the old
space_id is used.

The fix here is to add code into innobase_get_tablespace_statistics()
to use the current undo space ID. This is called during queries into
I_S.FILES for each space ID found by querying dd::tablespaces using
a read committed transaction.

There are a few extra changes which may be skipped, but that add a layer
of safety to internal InnoDB access to undo memory structures.

Approved by Marcin in RB#26219