BUG#32665990: DDL ON A SYSTEM TABLE COULD LEAD TO A SERVER               EXIT
ANALYSIS:
=========
It is advised not to tamper with any system tables.
Executing a DDL on a system table could result in unexpected
behavior when the server is restarted.

FIX:
====
An error is reported in such a case when the server is
restarted and the user is advised to run upgrade with FORCE
option.

Change-Id: Ibb2780ce319aaaac1daffabc5fc9bda4b8476381