Bug #32372038: SSL MEMORY LEAK IN THE VIO STRUCT ON CLEANUP RB#26216
The async code first creates an SSL context, stores it into the connect
async context of the MYSQL handle, then goes about doing SSL_connect() on
that.
But, I assume, this doesn't complete because of lack of bytes in the
buffer. So instead of transforming the vio into a SSL vio (and setting the SSL
context into it) it will just keep the SSL into the async connect context and
return "want more data" to the controlling application.
Now the controlling application most probably goes about doing mysql_close()
on that MYSQL structure. And, while doing that, it completely forgets
to free the SSL handle from the async connect context, if it's present.

Test case added sumilating the issue via the DBUG library.
Thanks to the mysql@facebook team for the contribution!