Bug #32169656: MYSQL GIVING QUERY RESULT FOR WRONG COLUMN
When resolving fields, the field name are compared with my_strcasecmp()
on utf8_general_ci (since it is system_charset_info). This does not follow
the general comparison rules of the collation; in particular, Ñ != N.
This is arguably wrong, but it is hard to change now, especially since
table creation follows the same rules.

However, when we have more than 32 fields in a table (this rule was set
before the initial VCS import), name lookup is done via a hash table.
Since hashing is collation-aware, it follows the correct rules, and
Ñ = N. This makes name lookup and duplication detection inconsistent.

We solve this by simply removing the hash; it is doubtful that it gives
any real performance gain.

Change-Id: I0e5ccf56b788869e538ef687b8173e41ded3cd8b