Bug #32558782: MALFORMED COMMUNICATION PACKET WITH BLOB COLUMN IN A PREPARED STATEMENT RB#26225
When implementing query attributes the format of the NET buffer for
COM_STMT_EXECUTE changed and the null bitmask is not at offset 0 in the
NET buffer anymore while serializing the data for the command.
The code was changed to store the char * to the actual location and pass
it down wherever needed to set the null bits. But this wasn't taking
into account that the NET could reallocate the buffer (and it does) if
the size is too small.
Fixed to not store the actual pointer, but offset from the start. And
now the null setting is using the actual NET buffer plus the offset it
received.
Test case added.