Bug#32831025 exit_code() called without wait_for_exit() [2/3]
Issue
=====

Currently the router only allows to signal:

- READY=1
- STOPPING

where READY=1 is after all components are "ready": metadata needs to
successfully fetch data from the metadata cluster.

Several tests check for failure scenarios which have to wait for the
plugins to be started, even if they'll never be READY.

There is a need for another sync-point once all components are running,
but not READY yet.

Change
======

- report "running" via sd_notify("STATUS=running");

RB: 26309