Bug #32644631: PRELIMINARY FIXES FOR WL #14419 [wf framebuffer, noclose] Bug #32802301: REGRESSION: TYPE_CONVERSION_STATUS FIELD_VARSTRING::STORE: ASSERTION `!TABLE || (!TABLE->WRITE_SET || BITMAP_IS_SET(TABLE->WRITE_SET, FIELD_INDEX()))' FAILED.
Some window functions need to look at overlapping portions of the
input rows, which in turn requires buffering rows in a special
framebuffer table. (It would probably be more efficient to use hash
join's row packing functionality, but it didn't exist when window
functions were written, and that is a fairly large rewrite.)

Currently, the windowing code has fairly complicated handling of
copies to and from the framebuffer's record buffers, stemming from
the fact that we set up the framebuffer before the output temporary
table (we conceptually always output to one, even though we might
not actually write the rows to disk). Ignoring the details,
fields (Item_field) are copied from the input (typically, but not
always, a single temporary table), while most functions are first
evaluated in the output temporary table before they are copied onwards
to the framebuffer.

This causes complexity in both setup (where the setup of the output
temporary table carefully needs to mark which functions are to be copied
on to the framebuffer), and also when bringing rows _back_ from the
framebuffer. In this case, some data will go to the output, but much
will also go to the input, necessitating an extra copy from the input to
the output.

We make two principal changes:

 - Set up the framebuffer after the output temporary table, not before.
   This allows us to always copy OUT -> FB and FB -> OUT; we never ever
   have IN -> FB and FB -> IN.

 - For window functions that we are to evaluate based on rows from the
   framebuffer (which is most buffering window functions), go through
   their arguments and replace fields in the input temporary table with
   the corresponding fields in the output temporary table. This allows
   us to evaluate them without an extra OUT -> IN copy.

This removes a number of special cases from temporary table setup,
and also fixes some oddities where the framebuffer would have
nonsensical temporary fields for the _output_ of window function
expressions, which would be copied back and forth.

We also restore some documentation on the buffering process that
was inadvertently removed with the pre-iterator executor; it has been
updated to reflect the current state.

The only output change in the test suite, save for EXPLAIN changes
and the likes, is an undefined case (setting of user variables do not
have well-defined evaluation times) in a “should not crash” test.

Change-Id: I7401e303efea23313bc6c481efe7bf86798abe2c