Bug #32458104  TEMPTABLE STORAGE ENGINE RETURNS ERROR: DOESN'T SUPPORT UPDATE OF INDEXES
Problem:

When using temptable storage engine in a release version (the setting is
internal_tmp_mem_storage_engine=TempTable), different errors would
return when aggregate functions are more than 255 in select list.

Analysis:

The class Indexed_column contains a member m_mysql_field_index, which is
supposed to hold the position of the indexed column (field that is part
of an index).  The data type used for this field was uint8_t.  So when
the position of the indexed field is > than 256, there is an overflow of
this variable.

Solution:

Change the datatype of Indexed_column::m_mysql_field_index so that it
can hold all possible field positions (1 to MAX_FIELDS).

rb#26230 in review by Jusufadis Bakamović <jusufadis.bakamovic@oracle.com>