Bug#32169848 THD_NDB TRANSACTION FUNCTIONALITY [#25]
Change to always register the handler for table stats in transaction and
remove the "local table stat instance" which is probably an old
optimization for simple autocommit with one involved hander.

This change removes several conditions and reduces memory consumption of
each involved ha_ndbcluster instance, instead one "emplace" into the
unordered_map will be done. If having a preallocated "table stat
instance" is better performance wise, it can be done as an
optimisation in Thd_ndb.

Change-Id: Id94f898341fd578d46a226db6f7b91928efa0974