Bug#32266286:ADDING GROUP BY SLOWS DOWN A QUERY
While evaluating loose index scan as a possible option
to do grouping and finding the min value, the cost
calculation does not reflect the fact that the query
will be looking at only one group because of the equality
predicates on the grouping attributes. This results in
looking at more rows as grouping is done after reading
the rows from the index.

Solution:
Determine if a query produces only one group, by
checking for the presence of equality predicates on
grouping attributes and use it for calculating
the cost. This will result in picking loose index scan
(if found beneficial) for such cases.

Change-Id: Ib2e2ba9ccb0b0c625045ee6d68a0419b3912425e