Bug#32523520: OPTIMIZER LOSES PART OF RANGE WHEN OR MULTIPLE               IDENTICAL RANGES IN RANGE PLAN
In a disjuctive predicate ((a=1 AND b=1 AND c=1)
OR (a=1 AND b=1 AND c=1) OR (a=1 AND b=2 AND c=2)),
while merging identical predicates on multiple keyparts of
an index (((a=1 AND b=1 AND c=1), (a=1 AND b=1 AND c=1)),
range optimizer does not remove the second predicate entirely.
Instead, it retains the predicate on just the first keypart.
((a=1 AND b=1 AND c=1) OR (a=1)). This results in re-examining
the same predicate only this time with no predicates on other
keyparts. Range optimizer further looks into second key parts
in the two predicates which are now (b=1) and NULL. This leads
dropping the range on second and further keyparts completely.
As a result, the query does not pick range scan later.

Solution:
Once the predicate is determined to be identical with another
predicate, after the range tree for the next key part is freed,
move to the next predicate in the range tree instead of still
pointing to partial predicate.

Patch contributed by Facebook.

Change-Id: Iec867237d6563a846438e9cd079da1a21e32e115