Bug#30789032: INCONSISTENT BEHAVIOR WHILE LOGGING A KILLED QUERY IN THE SLOW QUERY LOG
Problem with slow_query_log:
- with log_output=FILE
  - KILL QUERY:      logged
  - KILL CONNECTION: logged
- with log_output=TABLE
  - KILL QUERY:      logged
  - KILL CONNECTION: NOT logged

Where possible, the log destination should not affect
whether or not something is logged, as that is surprising.

At code level this is a result of logging to table
using open_table() which aborts if thd->killed is true.
Logging to file has no equivalent test, resulting in
asymmetrical behavior.

Solution
We already have a predicate for whether slow logging
should happen. By adding a test for KILL CONNECTION
to it, we achieve uniform behavior regardless of the
caller or the log destination. Thus, KILL QUERY will
result in a very long query being aborted, but still
logged for later investigation. KILL CONNECTION on
the other hand will result in the query being aborted,
the logging being precluded, and the connection being
dropped. This way, logging is no longer dependent on
an implementation detail. Instead, the user is afforded
additional control and a consistent experience, where
KILL QUERY will terminate just the query (but keep the
session, and as such, also the logging), and KILL
CONNECTION discards the entire session (query, logging,
and connection).

(Another solution would be to disable that test if one of
the submitted flags is MYSQL_LOCK_LOG_TABLE; yet another
to pass in MYSQL_OPEN_IGNORE_KILLED from open_log_table().
This would create a different kind of consistent behavior
where long queries are logged in all KILL scenarios.)

Admin note: Original patch was by Jens which as the
reviewer I inherited when he left the company.
I am therefore both the person approving the patch
and the one pushing it. -T-

Approved by: Tatiana Azundris Nuernberg <tatjana.nuernberg@oracle.com>

Change-Id: I15e8fb350a88c4142c271e2ab26dc8f1461466ee