Bug #32715698 UT0NEW ALLOCATOR IS NOT STL COMPLIANT
allocate() must never return a nullptr which ut_allocator under certain
circumstances can end up doing. This breaks the contract because standard
compliant allocators (like the ones which are planned to be used together
with STL containers) are always guaranteed to return a valid pointer.

Consequences of breaking this contract is such that STL container
implementation is not obliged to check for pointer validity, and that
will certainly lead to a non-obvious crash in STL which for sure won't
be trivial to debug. So, in the event of inability (*) to allocate memory,
standard-compliant allocators _must_ throw std::bad_alloc.

There's another (less of) violation and that one is that ut_allocator
will throw std::bad_alloc in case when there is available memory but
maximum number of elements of type T that can be fit into the array is
exceeded. Standard-compliant allocator in such cases must throw
std::bad_array_new_length.

Given that ut_allocator is used in number of places with STL containers
throughout the InnoDB code, this patch fixes above mentioned violations
by making the ut_allocator:
 (1) Throw std::bad_alloc in case it cannot fulfill the memory allocation
     request.
 (2) Throw std::bad_array_new_length in case max number of elements would
     have been exceeded.
 (3) Avoid returning a nullptr in case of 0-sized allocations

Last point turned out to be the most problematic one because making a
modification such that the ut_allocator throws an exception (or at least
debug asserts) in case of zero-sized allocation requests, it would cause
MySQL to crash with either unhandled exception or debug assert. This was
revealed thanks to InnoDB MTR test-suite and it is an indicator that
there _are_ several code paths which actually do end up doing the
unexpected zero-sized allocation requests.

Due to this side-effect this patch is only in position to remove the
zero-sized special handling all together because there are no other
viable options. It makes the ut_allocator such that the zero-sized
allocations will boil down to returning a result of malloc(0). Result of
malloc(0) is implementation-defined but it is a totally valid construct.
Pointer returned by it may typically refer to the zero-length block of
memory which _could_ consist of control data structures that underlying
system allocator implementation is making use of but it could also return
a nullptr. In latter case, ut_allocator will throw std::bad_alloc and
remain conformant and safe to use but in former case it will end up
triggering dangerous UB in which the application code is totally unaware
that it is dereferencing a non-nullptr that is not supposed to be
used at all. Modifying anything through that pointer will most likely
result with a crash in system allocator.

Given that zero-sized allocations are directly triggered by application
code, and there's nothing more that ut_allocator at this moment can do,
getting rid of them is out of the scope of this patch and will be
addressed through separate bug(s) because the fixes needed will require
a case-by-case evaluation.

(*) Inability to allocate memory is arguable and behavior varies from
platform to platform and obviously depends on respective virtual memory
settings. Without going into much details, this patch only tries to keep
the implementation sane and without any surprises.

Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB: 26272