Bug#29634540 "DROP DATABASE OF 1 MILLION TABLES RESULTED CRASH OF MYSQLD" [noclose]
Attempt to drop database with 1 million tables results in
out-of-memory condition.

With transition to New Data-Dictionary and atomic DDL, DROP DATABASE
started to allocate much more memory which is released at the end of
statement. This memory allocated for various DD API objects, new MDL
locks, temporary objects used during updates to DD tables and so on.

Part of the problem was that DD code responsible for updating SDI
in the general tablespace to which table being dropped belonged
created an uncached dd::Tablespace object for each table dropped
in order tto get tablespace name for acquiring MDL on it.
These objects were freed only at statement's end, even though they
were not used after acquiring MDL and there was no real need to keep
them for that long.

This patch solves this issue by changing code creating those uncached
dd::Tablespace objects to release them as soon as they become unnecessary.

There is no MTR test case as it is hard to write robust one for
this situation.

This patch is yet another small step towards solving the original
problem to the full extent.

Reviewed-by: Dyre Tjeldvoll <Dyre.Tjeldoll@oracle.com>