Bug #32790714: NEW FLAG PARAMETER_COUNT_AVAILABLE CAUSES CURSOR CREATION IN PREVIOUS SERVERS RB#26250
Made sending the flag optional on support of SEND_QUERY_METADATA
(8.0.23+) and the server version (from 8.0.25+).
This is because there is a bug in parsing the COM_STMT_EXECUTE flags
that conflates them all into a single boolean argument.
No change to the protocol description doxygen needed since it remains
the same, just that libmysql won't set that flag when talking to
8.0.23-8.0.24 servers.