Bug #32644631: PRELIMINARY FIXES FOR WL #14419 [aggressive walk, noclose]
When replacing parameters within window functions, don't limit ourselves
to Item_field for a given table; anything that is materialized is fair
game. This doesn't matter all that much for the old optimizer
(which typically materializes almost everything to fields before
windowing anyway), but the hypergraph optimizer needs it to get complex
expressions right.

This brings the function close enough to FindOrModifyReplacementItem()
that we can unify the two replacers into one.

Remove code from the dawn of time that would cause field names to be
different depending on whether -DNDEBUG was set, since that would leak
out into .result files now. (Item::item_name would be set in debug mode,
and not recomputed when that item later was materialized into a
temporary table field, even though it would get a slightly different
name at that time.)

Change-Id: Id09aaeb8ff18285fb26c96f0460a46777c4f2a94