BUG#31973053: UPGRADE CRASHES WITH FEDERATED SERVER CONNECTION
DESCRIPTION:
============
If a Federated table is created with server connection in
5.7 and the server upgrade to 8.0 is attempted, Upgrade
halts.

ANALYSIS:
=========
CREATE SERVER creates the definition of a server for use
with the FEDERATED storage engine. It also creates a new row
in the servers table in the mysql database.

When opening a federated table, we try to read a hashmap
called servers_cache to fetch the server name. This hashmap
is initialized by reading from the mysql.servers table. This
initialization was being done after the upgrade.

After upgrading a table, CHECK TABLE FOR UPGRADE is run for
the table. In the current scenario, CHECK TABLE for the
federated table tried to open the table and read the server
name from uninitialized hashmap. This lead to the server
exit.

FIX:
====
Server upgrade process involves first upgrading all tables
(system tables + user tables) and then run CHECK TABLE on
the upgraded tables. We are now initializing the structures
necessary for federated server after the upgrade before
CHECK TABLE is executed.

Change-Id: I4a0c4ef516c75ecc73d3b5a2a824ccecd5175b82