BUG#32554759 DO NOT ACCEPT -C OPTION WHEN STARTING WITH -F OR BINARY CONFIG
When mgmd is started with -f or config file the hostname and port
number on which the node should be started is mentioned in the config
file. -c or connectstring is used to fetch the configuration from
other mgmds. -c option is ignored when provided along with config-file.

This patch throws a warning when -c option is provided when starting
mgmd with config-file or binary config and continues its execution.

Change-Id: I169c69ad8d2ef67f88721e8d0a111c6763e73c0b