Bug#32713860: Executing prepare returns wrong result
maxmin_in_range() checks whether a condition applied to a row which
is read as the first or last row from an index is true.
Within this function there is a check that the condition is against
the field used when looking up the index. However, this check would
fail if the condition also contains a const-for-execution expression,
which it may do for dynamic parameters.

The fix is to replace the check with one that checks for presence of
the table, not equality.

There is also another problem: The function simple_pred() does not
describe fields that are compared to parameter values as "simple".
This means that predicates like "field OP parameter" do not qualify
for this optimization, unless the parameter has been converted to
a constant value earlier in the optimization. This happened for
parameters of type decimal, but not for type integer, thus the problem
does not happen for parameters of type integer. simple_pred() is also
adjusted to accept const-for-execution data as constant values,
along with being renamed to is_simple_predicate().

Also fixed some cast syntax and adjusted some comments.

Reviewed by: Chaithra Gopala Reddy <chaithra.gopalareddy@oracle.com>

Change-Id: Ib4c37202bbe1bbac5ee2227df5746d51bbd8608d