Bug #32778016 ASSERTION "K <= KMAX" FAILED|STRINGS/DTOA.CC
The function my_strtod_int() converts an input string into a
double. If the input string is very long, and contains mostly zeros,
we may hit an assert during conversion.

Solution: If we have scanned more than 308 digits after the '.' then
simply skip the rest of the input digits, and look for a possible
exponent part.

Change-Id: I54eaaf65e59c4f485d40c50f423f48781f638e20