Bug #32800130: CREATEBKAACCESSPATH() HIT ASSERT IN DOWN_CAST<STORE_KEY_FIELD *>
When we rewrite multi-equalities for BKA (because the access paths
might have put the tables in a different order from the old optimizer),
the source might be a field that comes from a multi-equality, but that
field may be in a const table. If so, the store_key is no longer a
store_key_field, but a store_key_const_item, so we cannot do the cast.

Longer-term, it would probably be good just to get rid of store_key_field
and use only store_key_item.

Change-Id: I898a6f3a8ef0e34c2390ad05ad0e36f435efc6d3