Bug #32715371 	INNODB: MAKE UT_LIST LENGTH MEMBER ATOMIC TO ALLOW LOCK-FREE QUERY WITHOUT UB
The ut_list requires an external protection to get current length of the list. In many instances we don't care for the value to be current, and we would like just to peek it.
Currently it is done in two ways:
1. With the external mutex acquired and momentarily released after reading the length.
2. Without the external mutex protection, this most probably works OK on x86 if compiler don't optimize it out, but still is an UB.

This change changes the ut_list length variable to be atomic, so the instances of 1st solution can be removed, and instances of 2nd are not
Undefined Behavior anymore. The operations on the variable are all made relaxed, as they are executed under the external protection and should not change the machine code generated for any such operation, at least on x86.

Additionally, the ut_list has the default constructor removed and a constructor with a pointer to field exposed. All list initialization is done now using the constructor.

RB#26172