WL#14628: Step 0: Move LOCK_replica_list to the global list
WL#14628: Compatibility option to enable old replication terminology
in P_S instrumentation

Context:

This is a simplification that was relevant in an earlier version of
the worklog.  Now the worklog does not depend on it, but we keep the
refactoring since it improves the code.

Background:

There is an API for registering mutex names with performance_schema.
Other server modules than replication don't invoke the API from the
module, but add the mutex to the list in mysqld.cc and rely on the
call to the API in mysqld.cc.  For other instrumentation classes than
mutexes - condition variables, memory allocations, thread stages,
thread names, etc - replication depends on the invocation in
mysqld.cc.  So it is only replication mutexes that have a special
case.

Problems:

- Replication mutexes are initialized in a different way, compared
  to other instrumentation, which is a bit inconsequent.
- There is a need for a specific call to a cleanup routine in
  replication.
- There is an error condition where slave_list_inited has not been
  initialized.  By initializing the mutex together with other mutexes,
  we ensure it happens very early during server start, so no code that
  uses the mutex can execute before that.

Fix:

Move the initialization of replication mutex instrumentation to the
global initialization.

RB:26366