Bug#32134145 - PARTITION PRUNING FAILS WITH EXPLICIT TIME_ZONE OFFSET IN DATETIME LITERAL
If a table is partitioned by timestamp and a timestamp literal with
timezone, is used in the where clause of a select statement a
partition may get omitted from the result set.

When timezone offset is specified in a timestamp literal, it is
expected to be converted to a timestamp without timezone after
adjusting the offset and then compared against a timestamp field. As
the literal is not adjusted, a partition can get pruned while
selecting the partitions to be scanned for the query.

Fix:
Make sure the timestamp with a timezone offset is converted before
comparing with the field values of the column.

Change-Id: Ie43305fa590d04c35bd886e92e7564fe4779cbb5