Bug #32695103 ASSERTION `RC == TYPE_OK' FAILED|SQL/DD/IMPL/RAW/RAW_RECORD.CC
The function make_unique_view_field_name() was documented to truncate
generated column names. However, it relied on snprintf to do the
truncation, and that does not work: It does not account for multibyte
characters, and the 'size' argument to snprintf changed from 64 to 192
at the time we switched to utf8mb3 as system character set.

Add code to count the number of characters of the generated name in
system character set, and truncate accordingly.

Change-Id: I95dbde4d3daea21e267ad60e0e1db076f551f7db