WL#14422: Full-text search in the hypergraph join optimizer [patch 9/12, limit hint]
In the cases where the optimizer has chosen a full-text index scan, it
is possible to give a limit as a hint to the storage engine, to signal
that it only needs to return a given number of matching documents.

This patch enables this limit hint in the index scan when the query
has a LIMIT clause, and it is known that there is no adding or
removing of rows before the limit operation happens. Specifically, it
enables the hint when all of these conditions are satisfied:

1) The query accesses a single table. No joins.

2) There is no aggregation and no DISTINCT.

3) There are no predicates, apart from those that can be subsumed by
the chosen index. (The subsumed predicates are those that evaluate to
TRUE for all rows returned by the index, and FALSE for all other rows,
so they don't filter out any of the rows returned by the full-text
index scan.)

Note that it is only a hint, and the storage engine may ignore it and
return more rows anyway, so we still need to add a LIMIT_OFFSET path
on top of the index scan. In particular, MyISAM always ignores the
hint, and InnoDB only uses the hint when it is given in combination
with the FT_NO_RANKING hint.

Change-Id: Iaf40c5fc0f061ef10ef4402405e0366063a69ce2