WL#14628: Step 2: New option that enables old terminology
WL#14628: Compatibility option to enable old replication terminology
in P_S instrumentation

Introduce the new configuration option
terminology_use_previous.  When set to NONE, the server is
not in compatibility mode and uses the newest version of all names.
When set to BEFORE_8_0_26, it uses the names that were present up to
and including 8.0.26.

We introduce a framework that translates new names to old names.  The
framework uses std::unordered_map objects where the keys are new names
and values are old names.  There is such an object for each
instrumentation class, i.e., one for mutexes, one for thread names,
etc. All those maps are collected in a std::unordered_map where the
keys are instrumentation class identifiers and values are the
maps. Finally, there is a vector where each element contains the
translation maps for a specific version.  This makes the framework
extensible, so in case we modify more names in the future, the
configuration option can be extended with a new enumeration value and
the vector with a new element.

We modify the class introduced in Step 1, so that:

- The setter for PFS_instr_name takes the instrumentation class as an
  argument.  For each version in the vector, it looks up the
  instrumentation class to get the translation map, and looks up the
  name in that map.  If the name is found, the setter caches the
  alternative name in member variables of PFS_instr_name.  It also
  caches the version in which the name appeared.

- The getter for PFS_instr_name checks if the session configuration
  option is enabled.  If it is, and the version specified in the option
  is older than the version cached in the member variable, it returns
  the alternative name cached in the member variable.

RB:26357