Bug #32831830 REGEXP TAKES MORE THAN 10X TIME WHEN COMPARED TO LIKE
This patch is not a bugfix, we merely provide microbenchmarks
comparing some examples of LIKE and RLIKE.

For some inputs, the ICU regex library is considerably slower than the
equivalend LIKE query. This patch adds some unit test benchmarks.

This patch is for 5.7. For comments, see the 8.0 patch.

Our BENCHMARK utilities in 8.0 require C++11, and have not been
backported (5.7 is C++03). Hence the timing code is slightly
different.

Change-Id: Idd1e2638dedfaede6caa9975ffb7ec9f80fb33c3