Bug#32855782 ROUTER LOG CONTENT IS PRINTED AS MOCK SERVERS LOG WHEN TESTS FAIL
When a component test fails the console output and log output
of each process gets dumped automatically.
The mock_server process currently does not log to the file but still
the Router process log content is dumped as a mock server log content.

This patch fixes this issue, Router log file is now assigned only
to the Router process.

RB: 26346