Bug#32855882 MYSQL FAILS TO ROLLBACK CREATE TABLE ... SELECT
Problem:
The system variable 'group_replication_transaction_size_limit' limits
maximum size of message for transaction which the replication group
accepts. When CREATE TABLE ... SELECT results in message exceeding this
limit (e.g. because it inserts too many bytes) the statement fails at
implicit commit time. The problem was that it also failed to rollback
properly.

The issue was caused by the fact that code in
Query_result_create::send_eof() call detected failure of implicit commit,
did some cleanup, but didn't propagate error up to its caller. As result
the caller behaved as is there was no error and called code which was not
ready for such situation.

This patch fixes the issue by propagating error from
Query_result_create::send_eof() to the caller in such cases. It also
simplifies code of this function by relying on
Query_result_create::abort_result_set() for cleanup in them.

Change-Id: I5c0bf2eb82579ce4f25fef7dee440787daf0cf7f